requirements.txt
Flask==3.0.0
pytesseract==0.3.10
Pillow==10.0.0
pdfplumber==0.10.3
spacy==3.7.2
python-docx==1.1.0
